+++ setup.py
@@ -12,7 +12,7 @@ setup(
         # We have type information
         'talepy': ['py.typed']
     },
-    version = "0.7.1",
+    version = "0.8.0",
     description = "Distributed Transactions Helper",
     author = "Steve Brazier",
     author_email = "meadsteve@gmail.com",
